+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.4.1 (2023-07-24 18:28)
+
+### Fixed
+
+* Fixed reconnect with pairing code or qrcode
+* Fixed problem in createJid
+
 # 1.4.0 (2023-07-24 17:03)
 
 ### Features